+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## [Unreleased]
 
- - 📝Improve the message from the crash when using a Spark component outside a `SparkTheme`
+ - 🐛 `ButtonContrast` and `ButtonGhost` were using `color` instead of `onContainerColor` values for their content colours
+ - 📝 Improve the message from the crash when using a Spark component outside a `SparkTheme`
 
 ## [1.8.0-alpha03]